refactor(execution): migrate autoMode.defaultAgent reads to resolveDefaultAgent (Phase 3B)
