+++ src/execution/unified-executor.ts
@@ -1,5 +1,6 @@
 /** Unified Story Executor (ADR-005, Phase 4) — sequential loop with optional parallel dispatch. */
 
+import { resolveDefaultAgent } from "../agents";
 import { checkCostExceeded, checkCostWarning, checkPreMerge, isTriggerEnabled } from "../interaction/triggers";
 import { getSafeLogger } from "../logger";
 import type { StoryMetrics } from "../metrics";
@@ -177,7 +178,7 @@ export async function executeUnified(
               story: { id: story.id, title: story.title, status: story.status, attempts: story.attempts },
               workdir: ctx.workdir,
               modelTier,
-              agent: ctx.config.autoMode.defaultAgent,
+              agent: ctx.agentManager?.getDefault() ?? resolveDefaultAgent(ctx.config),
               iteration: iterations,
             });
             logger?.info("story.start", `${story.title}`, {
@@ -279,7 +280,7 @@ export async function executeUnified(
               storyId: story.id,
               complexity: story.routing?.complexity ?? "medium",
               modelTier: story.routing?.modelTier ?? "balanced",
-              modelUsed: ctx.config.autoMode.defaultAgent,
+              modelUsed: ctx.agentManager?.getDefault() ?? resolveDefaultAgent(ctx.config),
               attempts: 1,
               finalTier: story.routing?.modelTier ?? "balanced",
               success: true,
@@ -301,7 +302,7 @@ export async function executeUnified(
                 storyId: conflict.story.id,
                 complexity: conflict.story.routing?.complexity ?? "medium",
                 modelTier: conflict.story.routing?.modelTier ?? "balanced",
-                modelUsed: ctx.config.autoMode.defaultAgent,
+                modelUsed: ctx.agentManager?.getDefault() ?? resolveDefaultAgent(ctx.config),
                 attempts: 1,
                 finalTier: conflict.story.routing?.modelTier ?? "balanced",
                 success: true,
@@ -373,7 +374,7 @@ export async function executeUnified(
             },
             workdir: ctx.workdir,
             modelTier,
-            agent: ctx.config.autoMode.defaultAgent,
+            agent: ctx.agentManager?.getDefault() ?? resolveDefaultAgent(ctx.config),
             iteration: iterations,
           });
           logger?.info("story.start", `${singleStory.title}`, {
@@ -464,7 +465,7 @@ export async function executeUnified(
         },
         workdir: ctx.workdir,
         modelTier,
-        agent: ctx.config.autoMode.defaultAgent,
+        agent: ctx.agentManager?.getDefault() ?? resolveDefaultAgent(ctx.config),
         iteration: iterations,
       });
       logger?.info("story.start", `${selection.story.title}`, {
